+++ XamarinCommunityToolkit.UnitTests/Xamarin.CommunityToolkit.UnitTests.csproj
@@ -20,6 +20,7 @@
   <ItemGroup>
     <Folder Include="Converters\" />
     <Folder Include="Behaviors\" />
+    <Folder Include="Mocks\" />
   </ItemGroup>
 
   <ItemGroup>
